+++ shared/src/index.ts
@@ -12,6 +12,7 @@ export * from "./probableLineup";
 export * from "./setPieceTaker";
 export * from "./manager";
 export * from "./user";
+export * from "./avatar";
 export * from "./purchase";
 export * from "./rosterExchange";
 export * from "./maxBid";